requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-google-genai
langchain-community
faiss-cpu
sentence-transformers
langchain
streamlit
langchain-groq

